CAMEL-20500: camel-jbang - Move generate commands to its own plugin
