+++ dsl/camel-jbang/camel-jbang-core/src/test/java/org/apache/camel/dsl/jbang/core/commands/plugin/PluginGetTest.java
@@ -69,11 +69,12 @@ class PluginGetTest extends CamelCommandBaseTest {
         command.doCall();
 
         List<String> output = printer.getLines();
-        Assertions.assertEquals(4, output.size());
+        Assertions.assertEquals(5, output.size());
         Assertions.assertEquals("Supported plugins:", output.get(0));
-        Assertions.assertEquals("NAME     COMMAND  DEPENDENCY                             DESCRIPTION", output.get(2));
+        Assertions.assertEquals("NAME      COMMAND   DEPENDENCY                                    DESCRIPTION", output.get(2));
         Assertions.assertEquals(
-                "camel-k  k        org.apache.camel:camel-jbang-plugin-k  %s".formatted(PluginType.CAMEL_K.getDescription()),
+                "camel-k   k         org.apache.camel:camel-jbang-plugin-k         %s"
+                        .formatted(PluginType.CAMEL_K.getDescription()),
                 output.get(3));
     }
 
@@ -117,17 +118,22 @@ class PluginGetTest extends CamelCommandBaseTest {
         command.doCall();
 
         List<String> output = printer.getLines();
-        Assertions.assertEquals(7, output.size());
+        Assertions.assertEquals(8, output.size());
         Assertions.assertEquals("NAME        COMMAND  DEPENDENCY                         DESCRIPTION", output.get(0));
         Assertions.assertEquals(
                 "foo-plugin  foo      org.apache.camel:foo-plugin:1.0.0  Plugin foo-plugin called with command foo",
                 output.get(1));
 
         Assertions.assertEquals("Supported plugins:", output.get(3));
-        Assertions.assertEquals("NAME     COMMAND  DEPENDENCY                             DESCRIPTION", output.get(5));
+        Assertions.assertEquals("NAME      COMMAND   DEPENDENCY                                    DESCRIPTION", output.get(5));
         Assertions.assertEquals(
-                "camel-k  k        org.apache.camel:camel-jbang-plugin-k  %s".formatted(PluginType.CAMEL_K.getDescription()),
+                "camel-k   k         org.apache.camel:camel-jbang-plugin-k         %s"
+                        .formatted(PluginType.CAMEL_K.getDescription()),
                 output.get(6));
+        Assertions.assertEquals(
+                "generate  generate  org.apache.camel:camel-jbang-plugin-generate  %s"
+                        .formatted(PluginType.GENERATE.getDescription()),
+                output.get(7));
     }
 
 }
