+++ dsl/camel-jbang/pom.xml
@@ -37,6 +37,7 @@
         <module>camel-jbang-console</module>
         <module>camel-jbang-core</module>
         <module>camel-jbang-main</module>
+        <module>camel-jbang-plugin-generate</module>
         <module>camel-jbang-plugin-k</module>
         <module>camel-jbang-it</module>
     </modules>
